requirements.txt
flask~=1.1.2
flask_sqlalchemy
flask_cors
Flask-Migrate
sqlalchemy~=1.3.17
psycopg2~=2.8.5
alembic~=1.4.2
python-jose